Move to demo to subpackage
